requirements.txt
streamlit
python-dotenv
langchain
openai
azure-storage-blob
PyMuPDF
python-docx
langchain-community